+++ app/src/main/java/th/in/route/routeinth/app/FirebaseUtils.java
@@ -30,7 +30,7 @@ public class FirebaseUtils {
         reference.child("users").child(uidUtils.getUID()).child("cardMap").child(c.getSystem()).setValue(c);
     }
 
-    public static void addValue(Context context, Card c, double value) {
+    public static void addValue(Context context, Card c, int value) {
         DatabaseReference reference = DatabaseUtils.getDatabase().getReference();
         UIDUtils uidUtils = new UIDUtils(context);
         reference.child("users").child(uidUtils.getUID()).child("cardMap").child(c.getSystem()).child("balance").setValue(c.getBalance() + value);
